fix: store dp as uint8 amplitude, not uint16 intensity
Tiled does not accept compressed write_block payloads (the server-side
CompressionMiddleware is response-only, and the only registered array
deserializers are application/octet-stream raw bytes plus a couple of
image formats). For local replay over WAN, that meant 8 MB per chunk ×
156 chunks = ~1.25 GB per scan, taking 10+ minutes to drain after the
75-second active publish phase.

Cut that in half by storing sqrt(intensity) rounded to uint8 instead of
raw uint16 intensity. The Eiger detector is 16-bit but sqrt(65535) ≈ 256,
so the full intensity range maps cleanly into uint8 with at most one
unit of clipping at the very top end. The 1-count quantization that
introduces is well below the Poisson noise floor for any pixel with
non-trivial counts, so it's lossless in practice for ML training.

ptycho-vit's Tiled-backed loader will skip its own sqrt step on this data
path (the bytes coming off Tiled are already amplitude). The h5 path is
unchanged.

Wire volume per 10K-frame 256×256 scan: 1.25 GB → 640 MB.

+++ holoptycho/ptycho_holo.py
@@ -622,10 +622,12 @@ class PtychoApp(Application):
         }
         _writer.start_run(self.run_uid, metadata=run_metadata)
         if enable_fine_tune_writes:
+            # uint8 storage is amplitude (sqrt of detector intensity). See
+            # TiledWriter.start_diffraction_buffer docstring for why.
             _writer.start_diffraction_buffer(
                 nz=int(x_num * y_num),
                 frame_shape=(int(self.param.nx), int(self.param.ny)),
-                dtype=np.uint16,
+                dtype=np.uint8,
             )
 
         # --- PtychoViT inference (parallel to iterative recon) ---
